obs_control/requirements.txt
discord.py
obs-websocket-py
